Changed file structure for easier debugging
